unit tests refactoring
